+++ trixnity-crypto-core/src/jsMain/kotlin/net/folivo/trixnity/crypto/core/aesCtr.kt
@@ -7,7 +7,6 @@ import js.typedarrays.Uint8Array
 import js.typedarrays.toUint8Array
 import kotlinx.coroutines.flow.flow
 import net.folivo.trixnity.utils.ByteArrayFlow
-import net.folivo.trixnity.utils.toByteArray
 import web.crypto.*
 
 actual fun ByteArrayFlow.encryptAes256Ctr(
@@ -15,7 +14,7 @@ actual fun ByteArrayFlow.encryptAes256Ctr(
     initialisationVector: ByteArray
 ): ByteArrayFlow {
     return if (PlatformUtils.IS_BROWSER) {
-        flow {// TODO should be streaming!
+        flow {
             val crypto = crypto.subtle
             val aesKey = crypto.importKey(
                 format = KeyFormat.raw,
@@ -24,18 +23,28 @@ actual fun ByteArrayFlow.encryptAes256Ctr(
                 extractable = false,
                 keyUsages = arrayOf(KeyUsage.encrypt, KeyUsage.decrypt),
             )
-            val result = Uint8Array(
-                crypto.encrypt(
-                    algorithm = jso<AesCtrParams> {
-                        name = "AES-CTR"
-                        counter = initialisationVector.toUint8Array()
-                        length = 64
-                    },
-                    key = aesKey,
-                    data = filterNotEmpty().toByteArray().toUint8Array(),
-                )
-            ).toByteArray()
-            emit(result)
+            val nonce = initialisationVector.copyOf(8)
+            var currentCounter = initialisationVector.copyOfRange(9, 16).toLong()
+            var previousInput = ByteArray(0)
+            filterNotEmpty().collect { nextInput ->
+                val input = previousInput + nextInput
+                val output = Uint8Array(
+                    crypto.encrypt(
+                        algorithm = jso<AesCtrParams> {
+                            name = "AES-CTR"
+                            counter = (nonce + currentCounter.toByteArray()).toUint8Array()
+                            length = 64
+                        },
+                        key = aesKey,
+                        data = input.toUint8Array(),
+                    )
+                ).toByteArray()
+                val nextOutput = output.copyOfRange(previousInput.size, output.size)
+                emit(nextOutput)
+                val inputSize = input.size / 16
+                currentCounter += inputSize
+                previousInput = input.copyOfRange(inputSize * 16, input.size)
+            }
         }.filterNotEmpty()
     } else {
         flow {
@@ -58,7 +67,7 @@ actual fun ByteArrayFlow.decryptAes256Ctr(
     initialisationVector: ByteArray
 ): ByteArrayFlow {
     return if (PlatformUtils.IS_BROWSER) {
-        flow {// TODO should be streaming!
+        flow {
             try {
                 val crypto = crypto.subtle
                 val aesKey = crypto.importKey(
@@ -68,18 +77,28 @@ actual fun ByteArrayFlow.decryptAes256Ctr(
                     extractable = false,
                     keyUsages = arrayOf(KeyUsage.encrypt, KeyUsage.decrypt),
                 )
-                val result = Uint8Array(
-                    crypto.decrypt(
-                        algorithm = jso<AesCtrParams> {
-                            name = "AES-CTR"
-                            counter = initialisationVector.toUint8Array()
-                            length = 64
-                        },
-                        key = aesKey,
-                        data = filterNotEmpty().toByteArray().toUint8Array(),
-                    )
-                ).toByteArray()
-                emit(result)
+                val nonce = initialisationVector.copyOf(8)
+                var currentCounter = initialisationVector.copyOfRange(9, 16).toLong()
+                var previousInput = ByteArray(0)
+                filterNotEmpty().collect { nextInput ->
+                    val input = previousInput + nextInput
+                    val output = Uint8Array(
+                        crypto.decrypt(
+                            algorithm = jso<AesCtrParams> {
+                                name = "AES-CTR"
+                                counter = (nonce + currentCounter.toByteArray()).toUint8Array()
+                                length = 64
+                            },
+                            key = aesKey,
+                            data = input.toUint8Array(),
+                        )
+                    ).toByteArray()
+                    val nextOutput = output.copyOfRange(previousInput.size, output.size)
+                    emit(nextOutput)
+                    val inputSize = input.size / 16
+                    currentCounter += inputSize
+                    previousInput = input.copyOfRange(inputSize * 16, input.size)
+                }
             } catch (exception: Throwable) {
                 throw AesDecryptionException(exception)
             }
@@ -102,4 +121,21 @@ actual fun ByteArrayFlow.decryptAes256Ctr(
             }
         }.filterNotEmpty()
     }
+}
+
+private fun ByteArray.toLong(): Long {
+    require(size <= 8)
+    var result = 0L
+    forEach { result = (result shl 8) + it }
+    return result
+}
+
+private fun Long.toByteArray(): ByteArray {
+    var intermediate = this
+    val result = ByteArray(8)
+    repeat(result.size) { i ->
+        result[i] = intermediate.toByte()
+        intermediate = intermediate shr 8
+    }
+    return result
 }