+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - fix download on missing file
 - never remove own keys from key tracking, when server says to
+- fix typo in encrypted file content
 
 ### Security